Move the ActionDialog component
