+++ frontend/components/ProgressSection.tsx
@@ -8,7 +8,7 @@ import ActionDataCollection from "@/data/actions.json"
 import FactionIcon from "@/components/FactionIcon"
 import { useGameContext } from "@/contexts/GameContext"
 import { useAuthContext } from "@/contexts/AuthContext"
-import ActionDialog from "@/components/actionDialogs/ActionDialog"
+import ActionDialog from "@/components/ActionDialog"
 import ActionDataCollectionType from "@/types/Action"
 import Faction from "@/classes/Faction"
 import FactionLink from "@/components/FactionLink"
